Integrate Supabase code with Vue composables.
1. Auth Composable
`useAuth.ts`
A dedicated composable for sign up / sign in / sign out.

2. Avatar Composable
`useProfile.ts`
Place getProfile() and updateProfile() in a composable.

3. Profile Composable
 `useAvatar.ts`
 Create a separate composable for
` downloadImage()` & `uploadAvatar()`
