+++ package.json
@@ -19,9 +19,10 @@
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.16",
     "@vitejs/plugin-vue": "^5.2.1",
     "@vue/tsconfig": "^0.7.0",
-    "typescript": "~5.6.2",
+    "typescript": "~5.7.3",
     "vite": "^6.0.11",
     "vue-tsc": "^2.2.0"
   }
