+++ src/components/Auth.vue
@@ -1,43 +1,27 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import { supabase } from '../supabase'
-import { useRouter } from 'vue-router'
+import { useAuth } from '../composables/useAuth'
 
-const router = useRouter()
 const email = ref('')
 const password = ref('')
-const loading = ref(false)
+const { loading, errorMsg, signIn } = useAuth()
 
-async function signIn() {
-  try {
-    loading.value = true
-    const { error } = await supabase.auth.signInWithPassword({
-      email: email.value,
-      password: password.value
-    })
-    if (error) throw error
-    router.push('/second')
-
-  } catch (error : any) {
-    alert(error.message)
-  } finally {
-    loading.value = false
-  }
+function handleSignIn() {
+  signIn(email.value, password.value)
 }
 </script>
 
 <template>
   <div class="max-w-md mx-auto">
     <div class="mb-4">
-      <label for="email" class="block text-sm font-medium text-gray-700">Email</label>
-      <input id="email" type="email" v-model="email" class="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
+      <label class="block text-sm font-medium text-gray-700">Email</label>
+      <input type="email" v-model="email" class="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
     </div>
     <div class="mb-4">
-      <label for="password" class="block text-sm font-medium text-gray-700">Password</label>
-      <input id="password" type="password" v-model="password" class="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
+      <label  class="block text-sm font-medium text-gray-700">Password</label>
+      <input type="password" v-model="password" class="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
     </div>
-    <div>
-      <button class="button block w-full px-4 py-2 bg-indigo-600 text-white font-semibold rounded-md shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500" @click="signIn" :disabled="loading">Sign In</button>
-    </div>
+      <button class="button block w-full px-4 py-2 bg-indigo-600 text-white font-semibold rounded-md shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500" @click="handleSignIn" :disabled="loading">Sign In</button>
+        <p v-if="errorMsg" class="error">{{ errorMsg }}</p>
   </div>
 </template>
